requirements.txt
requests
beautifulsoup4
supabase
python-dotenv
python-slugify
cloudscraper
brotli
yt-dlp